- Move `abEnvToCfg`. - Remove compiler warnings.
