+++ test/Test/Cardano/Chain/Block/Elaboration.hs
@@ -1,61 +1,62 @@
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
-{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE OverloadedLists #-}
 
 -- | This module provides functionality for translating abstract blocks into
 -- concrete blocks. The abstract blocks are generated according the small-step
 -- rules for the block chain (also called the blockchain specification).
 module Test.Cardano.Chain.Block.Elaboration
-  ( elaborate
+  ( abEnvToCfg
+  , elaborate
   , elaborateBS
   )
 where
 
 import Cardano.Prelude hiding (to)
 
-import Control.Lens ((^.), to, makeLenses, (.~), (^..))
+import Control.Lens ((^.), to, (^..))
 import qualified Data.ByteString.Lazy as LBS
 import Data.Coerce (coerce)
-import qualified Data.Map.Strict as M
+import qualified Data.Map.Strict as Map
+import qualified Data.Set as Set
+import Data.Time (Day(ModifiedJulianDay), UTCTime(UTCTime))
 
 import qualified Cardano.Binary.Class as Binary
-import qualified Test.Cardano.Crypto.Dummy as Crypto
 import qualified Cardano.Crypto.Hashing as H
-import qualified Cardano.Crypto.Signing as Signing
-import qualified Cardano.Crypto.Wallet as CC
 
 import qualified Cardano.Chain.Block as Concrete
 import qualified Cardano.Chain.Common as Common
 import qualified Cardano.Chain.Delegation as Delegation
 import qualified Cardano.Chain.Genesis as Genesis
 import qualified Cardano.Chain.Ssc as Ssc
-import Cardano.Chain.Slotting (SlotId (SlotId))
-import qualified Cardano.Chain.Slotting as Slotting
 import qualified Cardano.Chain.Txp as Txp
 import qualified Cardano.Chain.Update as Update
+import qualified Cardano.Chain.Slotting as Slotting
 
 import qualified Control.State.Transition as Transition
-import Cardano.Spec.Chain.STS.Rule.Chain (CHAIN, disL, epochL, ppsL)
+import Cardano.Spec.Chain.STS.Rule.Chain (CHAIN, disL, epochL)
 import qualified Cardano.Spec.Chain.STS.Block as Abstract
 import qualified Ledger.Core as Abstract
 import Ledger.Delegation (DCert, delegationMap, delegatorOf, mkDCert)
-import Ledger.Update (bkSlotsPerEpoch, PParams)
+import Ledger.Update (bkSgnCntW, bkSlotsPerEpoch, maxBkSz, maxHdrSz)
+import Cardano.Chain.Common
+  ( BlockCount(BlockCount)
+  , LovelacePortion(LovelacePortion)
+  , TxFeePolicy(TxFeePolicyUnknown)
+  , mkStakeholderId
+  )
 
+import Test.Cardano.Crypto.Dummy (dummyProtocolMagic)
 import Test.Cardano.Chain.Interpreter
-  ( interpretDCert
+  ( interpretDCert, elaborateVKeyGenesis
   , interpretKeyPair
   , vKeyPair
   )
 
--- TODO: discuss with Ru whether the genesis hash can be in the
--- 'ChainValidationState'.
+-- | Elaborate an abstract block into a concrete block (without annotations).
 elaborate
-  :: Genesis.Config -- TODO: Do we want this to coincide with the hash of the
-                    -- abstract environment? (and in such case we wouldn't need
-                    -- this parameter)
+  :: Genesis.Config
   -> Transition.Environment CHAIN
   -> Transition.State CHAIN
   -> Concrete.ChainValidationState
@@ -72,11 +73,10 @@ elaborate config (_, _, pps) ast st ab
       = Concrete.mkHeaderExplicit
         (Genesis.configProtocolMagicId config)
         prevHash
-        0 -- Chain difficulty.
-        -- TODO: we might want to generate the chain difficulty
-        sid -- SlotId
-        ssk -- SecretKey
-        cDCert -- Maybe Delegation.Certificate
+        0
+        sid
+        ssk
+        cDCert
         bb0
         extraHeaderData
 
@@ -96,14 +96,12 @@ elaborate config (_, _, pps) ast st ab
       }
 
     prevHash
-      = fromMaybe (Genesis.getGenesisHash $ Genesis.configGenesisHash config) $ Concrete.cvsPreviousHash st
+      = fromMaybe (Genesis.getGenesisHash $ Genesis.configGenesisHash config)
+                  (Concrete.cvsPreviousHash st)
 
-    sid = Slotting.unflattenSlotId
-      (coerce (pps ^. bkSlotsPerEpoch))
-      -- TODO: I don't like this inconsistency between qualifying and not, but
-      -- some qualifying some identifiers can get too awkward. And using this
-      -- with lenses gets even worse.
-      (ab ^. Abstract.bHeader . Abstract.bSlot . to Abstract.unSlot)
+    sid
+      = Slotting.unflattenSlotId (coerce (pps ^. bkSlotsPerEpoch))
+                                 (ab ^. Abstract.bHeader . Abstract.bSlot . to Abstract.unSlot)
 
     issuer = ab ^. Abstract.bHeader . Abstract.bIssuer
 
@@ -138,16 +136,19 @@ annotateBlock
   :: Concrete.Block
   -> Concrete.ABlock ByteString
 annotateBlock block =
-  -- TODO: use 'encode' from 'Bi'
   let res =
         case Binary.decodeFullDecoder "Block" Concrete.decodeABlockOrBoundary bytes of
           Left err ->
-            panic $ "This function should be able to decode the block it encoded. Instead I got: " <> show err
+            panic $ "This function should be able to decode the block it encoded"
+                  <> ". Instead I got: " <> show err
           Right abobb ->
             map (LBS.toStrict . Binary.slice bytes) abobb
   in
   case res of
-    Concrete.ABOBBlock bk -> bk
+    Concrete.ABOBBlock bk   ->
+      bk
+    Concrete.ABOBBoundary _ ->
+      panic "This function should have decoded a block."
 
   where
     bytes = Binary.serializeEncoding (Concrete.encodeBlock block)
@@ -169,8 +170,73 @@ rcDCert vk ast
 
     vkg = fromMaybe err $ delegatorOf dm vk
 
+    err :: Abstract.VKeyGenesis
     err = panic $ "No delegator found for key " <> show vk
 
     vkp = vKeyPair $ coerce vkg
 
     sigVkg = Abstract.sign (Abstract.sKey vkp) vkg
+
+--  | Make a genesis configuration from an initial abstract environment of the
+--  | trace.
+--
+abEnvToCfg
+  :: Transition.Environment CHAIN
+  -> Genesis.Config
+abEnvToCfg (_, vkgs, pps) = Genesis.Config genesisData genesisHash Nothing
+  where
+    genesisData
+      = Genesis.GenesisData
+      { Genesis.gdBootStakeholders =
+          Genesis.GenesisWStakeholders genesisStakeHolders
+      , Genesis.gdHeavyDelegation =
+          Genesis.UnsafeGenesisDelegation [] -- We don't need initial heavyweight delegation.
+      , Genesis.gdStartTime =
+          UTCTime (ModifiedJulianDay 0) 0
+      , Genesis.gdNonAvvmBalances =
+          Genesis.GenesisNonAvvmBalances []
+      , Genesis.gdProtocolParameters =
+          gPps
+      , Genesis.gdK =
+          -- TODO: this should be a different protocol parameter once we have
+          -- an abstract protocol parameter for k. Then we need to solve the
+          -- problem that in the concrete implementation k and w are the same.
+          BlockCount  (fromIntegral $ pps ^. bkSgnCntW)
+      , Genesis.gdProtocolMagic =
+          dummyProtocolMagic
+      , Genesis.gdAvvmDistr =
+          Genesis.GenesisAvvmBalances []
+      }
+
+    -- We shouldn't need to use 'coerce' after
+    -- https://github.com/input-output-hk/cardano-chain/issues/332 gets
+    -- implemented.
+    genesisHash = Genesis.GenesisHash $ coerce $ H.hash ("" :: ByteString)
+
+    gPps
+      = Update.ProtocolParameters
+      { Update.ppScriptVersion = 0
+      , Update.ppSlotDuration = 0
+      , Update.ppMaxBlockSize = pps ^. maxBkSz
+      , Update.ppMaxHeaderSize = pps ^. maxHdrSz
+      , Update.ppMaxTxSize = 0
+      , Update.ppMaxProposalSize = 0
+      , Update.ppMpcThd = LovelacePortion 0
+      , Update.ppHeavyDelThd = LovelacePortion 0
+      , Update.ppUpdateVoteThd = LovelacePortion 0
+      , Update.ppUpdateProposalThd = LovelacePortion 0
+      , Update.ppUpdateImplicit = 0
+      , Update.ppSoftforkRule =
+          Update.SoftforkRule (LovelacePortion 0) (LovelacePortion 0) (LovelacePortion 0)
+      , Update.ppTxFeePolicy =
+          TxFeePolicyUnknown 0 "Fee policy unspecified at \"Test.Cardano.Chain.Block.Validation.Spec\""
+      , Update.ppUnlockStakeEpoch = 0
+      }
+
+
+    genesisStakeHolders :: Map Common.StakeholderId Word16
+    genesisStakeHolders
+      = Map.fromList
+      $ zip (mkStakeholderId . elaborateVKeyGenesis <$> vkgs') [1..]
+
+    vkgs' = Set.toList vkgs
